+++ negotiation/src/main/java/it/eng/negotiation/model/Agreement.java
@@ -7,7 +7,6 @@ import java.util.stream.Collectors;
 
 import org.springframework.data.mongodb.core.mapping.Document;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
@@ -30,11 +29,6 @@ import lombok.NoArgsConstructor;
 @Document(collection = "agreements")
 public class Agreement {
 
-	@JsonIgnore
-	private String consumerPid;
-	
-	@JsonIgnore
-	private String providerPid;
 	/*
 	"dspace:agreement": {
     "@id": "urn:uuid:e8dc8655-44c2-46ef-b701-4cffdc2faa44",
@@ -109,16 +103,6 @@ public class Agreement {
 			return this;
 		}
 		
-		public Builder consumerPid(String consumerPid) {
-			agreement.consumerPid = consumerPid;
-			return this;
-		}
-		
-		public Builder providerPid(String providerPid) {
-			agreement.providerPid = providerPid;
-			return this;
-		}
-
 		@JsonProperty(DSpaceConstants.ODRL_TARGET)
 		public Builder target(String target) {
 			agreement.target = target;