+++ negotiation/src/test/java/it/eng/negotiation/service/ContractNegotiationAPIServiceTest.java
@@ -37,6 +37,7 @@ import it.eng.negotiation.repository.ContractNegotiationRepository;
 import it.eng.negotiation.repository.OfferRepository;
 import it.eng.negotiation.serializer.Serializer;
 import it.eng.tools.client.rest.OkHttpRestClient;
+import it.eng.tools.model.IConstants;
 import it.eng.tools.response.GenericApiResponse;
 import it.eng.tools.util.CredentialUtils;
 
@@ -62,6 +63,8 @@ public class ContractNegotiationAPIServiceTest {
 	private ArgumentCaptor<ContractNegotiation> argCaptorContractNegotiation;
 	@Captor
 	private ArgumentCaptor<Agreement> argCaptorAgreement;
+	@Captor
+	private ArgumentCaptor<Offer> argCaptorOffer;
 
 	@InjectMocks
 	private ContractNegotiationAPIService service;
@@ -74,11 +77,14 @@ public class ContractNegotiationAPIServiceTest {
 		when(apiResponse.getData()).thenReturn(Serializer.serializeProtocol(MockObjectUtil.CONTRACT_NEGOTIATION_ACCEPTED));
 		when(apiResponse.isSuccess()).thenReturn(true);
 		when(properties.consumerCallbackAddress()).thenReturn(MockObjectUtil.CALLBACK_ADDRESS);
+		when(offerRepository.save(any(Offer.class))).thenReturn(MockObjectUtil.OFFER_WITH_ORIGINAL_ID);
 		
 		service.startNegotiation(MockObjectUtil.FORWARD_TO, Serializer.serializePlainJsonNode(MockObjectUtil.OFFER));
 		
-		verify(contractNegotiationRepository).save(any(ContractNegotiation.class));
-		verify(offerRepository).save(any(Offer.class));
+		verify(offerRepository).save(argCaptorOffer.capture());
+		verify(contractNegotiationRepository).save(argCaptorContractNegotiation.capture());
+		assertEquals(IConstants.ROLE_CONSUMER, argCaptorContractNegotiation.getValue().getRole());
+		assertEquals(MockObjectUtil.OFFER.getId(), argCaptorOffer.getValue().getOriginalId());
 	}
 	
 	@Test
@@ -276,9 +282,33 @@ public class ContractNegotiationAPIServiceTest {
 				.thenReturn(Arrays.asList(
 						MockObjectUtil.CONTRACT_NEGOTIATION_ACCEPTED,
 						MockObjectUtil.CONTRACT_NEGOTIATION_REQUESTED));
-		Collection<JsonNode> response = service.findContractNegotiations(null, null);
+		Collection<JsonNode> response = service.findContractNegotiations(null, null, null);
 		assertNotNull(response);
 		assertEquals(2, response.size());
+	}
+	
+	@Test
+	@DisplayName("Find contract negotiations by role")
+	public void findContractNegotiationByRole() {
+		when(contractNegotiationRepository.findAll())
+				.thenReturn(Arrays.asList(
+						MockObjectUtil.CONTRACT_NEGOTIATION_ACCEPTED,
+						MockObjectUtil.CONTRACT_NEGOTIATION_REQUESTED));
+		Collection<JsonNode> response = service.findContractNegotiations(null, null, IConstants.ROLE_CONSUMER);
+		assertNotNull(response);
+		assertEquals(2, response.size());
+	}
+	
+	@Test
+	@DisplayName("Find contract negotiations by role - not found")
+	public void findContractNegotiationByRole_other() {
+		when(contractNegotiationRepository.findAll())
+				.thenReturn(Arrays.asList(
+						MockObjectUtil.CONTRACT_NEGOTIATION_ACCEPTED,
+						MockObjectUtil.CONTRACT_NEGOTIATION_REQUESTED));
+		Collection<JsonNode> response = service.findContractNegotiations(null, null, IConstants.ROLE_PROVIDER);
+		assertNotNull(response);
+		assertEquals(0, response.size());
 	}
 	
 	@Test
@@ -286,7 +316,7 @@ public class ContractNegotiationAPIServiceTest {
 	public void findContractNegotiationById() {
 		when(contractNegotiationRepository.findById(MockObjectUtil.CONTRACT_NEGOTIATION_ACCEPTED.getId()))
 				.thenReturn(Optional.of(MockObjectUtil.CONTRACT_NEGOTIATION_ACCEPTED));
-		Collection<JsonNode> response = service.findContractNegotiations(MockObjectUtil.CONTRACT_NEGOTIATION_ACCEPTED.getId(), null);
+		Collection<JsonNode> response = service.findContractNegotiations(MockObjectUtil.CONTRACT_NEGOTIATION_ACCEPTED.getId(), null, null);
 		assertNotNull(response);
 		assertEquals(1, response.size());
 	}
@@ -296,7 +326,7 @@ public class ContractNegotiationAPIServiceTest {
 	public void findContractNegotiationByState() {
 		when(contractNegotiationRepository.findByState(ContractNegotiationState.ACCEPTED.name()))
 				.thenReturn(Arrays.asList(MockObjectUtil.CONTRACT_NEGOTIATION_ACCEPTED));
-		Collection<JsonNode> response = service.findContractNegotiations(null, ContractNegotiationState.ACCEPTED.name());
+		Collection<JsonNode> response = service.findContractNegotiations(null, ContractNegotiationState.ACCEPTED.name(), null);
 		assertNotNull(response);
 		assertEquals(1, response.size());
 	}
@@ -352,10 +382,12 @@ public class ContractNegotiationAPIServiceTest {
 		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class))).thenReturn(apiResponse);
 		when(apiResponse.isSuccess()).thenReturn(true);
 		
-		service.handleContractNegotiationAgreed(contractNegotaitionId);
+		service.approveContractNegotiation(contractNegotaitionId);
 		
+		verify(agreementRepository).save(argCaptorAgreement.capture());
 		verify(contractNegotiationRepository).save(argCaptorContractNegotiation.capture());
 		assertEquals(ContractNegotiationState.AGREED, argCaptorContractNegotiation.getValue().getState());
+		assertEquals(argCaptorAgreement.getValue().getId(), argCaptorContractNegotiation.getValue().getAgreement().getId());
 	}
 	
 	@Test
@@ -365,7 +397,7 @@ public class ContractNegotiationAPIServiceTest {
 		when(contractNegotiationRepository.findById(contractNegotaitionId)).thenReturn(Optional.of(MockObjectUtil.CONTRACT_NEGOTIATION_AGREED));
 		
 		assertThrows(ContractNegotiationAPIException.class, 
-				() -> service.handleContractNegotiationAgreed(contractNegotaitionId));
+				() -> service.approveContractNegotiation(contractNegotaitionId));
 	}
 	
 	@Test
@@ -380,7 +412,7 @@ public class ContractNegotiationAPIServiceTest {
 		when(apiResponse.isSuccess()).thenReturn(false);
 		
 		assertThrows(ContractNegotiationAPIException.class, 
-				() -> service.handleContractNegotiationAgreed(contractNegotaitionId));
+				() -> service.approveContractNegotiation(contractNegotaitionId));
 		
 		verify(contractNegotiationRepository, times(0)).save(argCaptorContractNegotiation.capture());
 		verify(agreementRepository, times(0)).save(argCaptorAgreement.capture());